Marching squares drawn
Have an implementation that draws marching squares.  User can click to
toggle the state of a cell corner and the squares are updated.
This is all drawn into a container, so some methods need created to
interact with the boundaries.  Collision detection and such.
Also need to play with randomizing all of the corners to see how that
looks.
Need to create a means of toggling all the corners by interacting with a
cell.