Adding base content
Updated the container class so that it can be resized after creation.

Added card images for the automated overlord deck

Still need to determine the layout for the Descent App so that the UI
controls can be created.  Added a function to load the card images based
on an XML configuration file.