Getting cards on screen
The draw and display of the automated overlord cards is working now.
Need to add in something for the reference cards still.

Added a frame for the overlord cards, duplicated automated cards.  This
isn't ideal since a hand needs to be displayed not just one card, but
it's code duplication.  Better solution needed.

Added overlord card inages into the content directory.  Just a basic
set, not sure how many will be needed.