Card Decks working
Have a common control to handle the card decks.  It will display draw
cards face up.  Click to zoom and will trigger events when a card is
clicked on.

There is an option to remove cards as well as shuffle them.

The reference cards are not currently loaded, so that needs added again.
But with no container to display them its not a big deal.

Next to get the character sheets built.