Updated tools and sprite manager
Merged in a number of basic C# utility functions for working with Enums.
A new class space specific to enums was created to contain them.

An initial sprite manager is now included as well.  It loads tile
information and animation sequences from an XML file.  It is able to
draw the sprites and carry out basic animations.