Tilesets and asteroid mechanics
Added a series of classes to manage and display tile sets along with
basic animation of sprites.

Corrected the transparency handling of the container class.

Also Added some mechanics and advancements i nthe asteroid shooter game.
Things like errors in aiming of the hunters, gun inprovements with kills
for the player and different rates of asteroid spawns.  Makes it more
interesting, but still incomplete as a game.  Good study of improving
mechanics though.