Filled polygons
Added code to draw primitive shapes through the graphics device.  Fillign in a series of triangles allowed the polygons to be filled in.
There's some oddity in how the list of triangles are being filled in.  I think it is doing some overdraw to accomplish this.  There should only be 2 triangles needed to fill it, but for some reason 3 are necessary.

Also when using 4 vertexes in the polygons some false intersections are being reported.