Fixed polygon fill
Fixed the logic in the point in triangle function so that it doesn't give false positive results.
Corrected the primitive drawing in the polygon class so that the minimum number of triangles are used.
Ran tests using various vertex counts in the polygons and everything seems to be working correctly.