Building physical object class
Creating some initial classes to provide basic functionality for the objects that will be drawn to the screen.
Right now I'm trying to ensure that when I move the objects around all of the other details move with them.
The tecture appears to work, but the collision polygon needs some additional controls.
Also the convex polygon class no longer seems to be able to fill the polygon.