Can't Mix Draw modes
Found out one of my Shader properties was incorrect.  Polygons can be filled again.

You can't do direct graphics device draws in the middle of sprite batch draws.
The sprite batches are rendered when End() is called, so any graphics device draws will be clobbered when that happens.
Since ConvexPolygon is the only class that does this it will handle all of its drawing and will not accept a SpriteBatch parameter