PhysicalObject Drawing
The SpriteBatch.Drw() scales the image first, then rotates it.  Which means it will rotate at the correct position, but will fall outside the draw region specified.
This probably isn't wrong, but it doesn't always follow intuition.  If the scalign is close to 1:1 for X:Y then you never notice.