Visual manipulation fixes
Had to remove the image offset, that was throwing off a lot of the calculations since pieces were not rotated equally.
After doing that making the scaling functions work correctly was fairly easy.
This should complete the physical object class and allow for quick creation of objects with decent functionality.