Class updates
Modified the particle engine to use an interface for what it manages.  This allows easier importing of external objects.
The object manager can now apply the loaded definitions to external objects and not manage them.  Also added default Height/Width values to the definitions since they can't be changed without impacting the collision polygon.
Added a Missile extension of the Physical object which can be juggled by the particle engine  to be player shots.