Trackign missiles
Came up with an algorithm for missiles to chase after a target.
This makes use of the interfaces built before and handles all the game object through existing systems.

There's no connection between ObjectManager groups and the Particle engine which leaves a bit of an information gap with goruping objects.  Maybe th emanager needs to also do the particle work?