New libraries
Built a new texture atlas class that draws the tiles using primitives to allow for more flexibility when putting the textures on the screen.  Full polygon support instead of just rectangles.  This feels a lot like a new version of the sprite batch object.

Added some matrix math functions for doing vertex transformations.  Not sure these are faster, but they give central place to update the transforms so they'll be common from object to object.